feat: add backward compatibility
