+++ src/app/docs/route.ts
@@ -1,0 +1,5 @@
+import { redirect } from 'next/navigation'
+
+export async function GET() {
+  redirect('/doc/')
+}
